# src/main/resources/swagger/server/api.json
{
  "swagger": "2.0",
  "info": {
    "title": "E2D masterdata service",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/instructor/get/{id}": {
      "get": {
        "tags": [
          "instructor"
        ],
        "summary": "Get instructor",
        "operationId": "getInstructor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format":"int64"
          }
        ],
        "responses": {
          "200": {
            "description": "instructor object",
            "schema": {
              "$ref": "#/definitions/Instructor"
            }
          }
        }
      }
    },
...
    "/instructor/update/{id}": {
      "put": {
        "tags": [
          "instructor"
        ],
        "summary": "update instructor",
        "operationId": "updateInstructor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format":"int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Instructor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "instructor object",
            "schema": {
              "$ref": "#/definitions/Instructor"
            }
          }
        }
      }
    },
    "/instructor/delete/{id}": {
      "delete": {
        "tags": [
          "instructor"
        ],
        "summary": "Delete instructor",
        "operationId": "deleteInstructor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format":"int64"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/school/get/{id}": {
      "get": {
        "tags": [
          "school"
        ],
        "summary": "Get school",
        "operationId": "getSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format":"int64"
          }
        ],
        "responses": {
          "200": {
            "description": "school object",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Instructor": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "user":  {
          "$ref": "#/definitions/User"
        },
        "schoolId": {
          "type": "integer",
          "format":"int64"

        }
      }
    },
...
    "School": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "user":  {
          "$ref": "#/definitions/User"
        },
        "name": {
          "type": "string"
        },
        "officialName": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "firstName": {
          "type": "string"
        },

        "secondName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    }
  },
...
}